+++ leeedohyun/src/main/java/movie/AmountDiscountPolicy.java
@@ -1,6 +1,6 @@
 package movie;
 
-public class AmountDiscountPolicy extends DiscountPolicy {
+public class AmountDiscountPolicy extends DefaultDiscountPolicy {
 
     private Money discountAmount;
