+++ leeedohyun/src/main/java/movie/DiscountPolicy.java
@@ -1,25 +1,6 @@
 package movie;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+public interface DiscountPolicy {
 
-public abstract class DiscountPolicy {
-
-    private List<DisCountCondition> conditions = new ArrayList<>();
-
-    public DiscountPolicy(DisCountCondition... conditions) {
-        this.conditions = Arrays.asList(conditions);
-    }
-
-    public Money calculateDiscountAmount(Screening screening) {
-        for (DisCountCondition each : conditions) {
-            if (each.isSatisfiedBy(screening)) {
-                return getDiscountAmount(screening);
-            }
-        }
-        return Money.ZERO;
-    }
-
-    abstract protected Money getDiscountAmount(Screening screening);
+    Money calculateDiscountAmount(Screening screening);
 }
