+++ leeedohyun/src/main/java/movie/PercentDiscountPolicy.java
@@ -1,6 +1,6 @@
 package movie;
 
-public class PercentDiscountPolicy extends DiscountPolicy {
+public class PercentDiscountPolicy extends DefaultDiscountPolicy {
 
     private double percent;
