+++ leeedohyun/src/test/java/movie/DiscountPolicyTest.java
@@ -13,7 +13,7 @@ class DiscountPolicyTest {
     void 금액_할인_정책을_적용한다() {
         // given
         SequenceCondition sequenceCondition = new SequenceCondition(10);
-        DiscountPolicy discountPolicy = new AmountDiscountPolicy(Money.wons(1000), sequenceCondition);
+        DefaultDiscountPolicy discountPolicy = new AmountDiscountPolicy(Money.wons(1000), sequenceCondition);
         Movie movie = new Movie("아바타", Duration.ofMinutes(100), Money.wons(10000), discountPolicy);
         Screening screening = new Screening(movie, 10, LocalDateTime.of(2025, 4, 21, 10, 0));
 
@@ -28,7 +28,7 @@ class DiscountPolicyTest {
     void 비율_할인_정책을_적용한다() {
         // given
         SequenceCondition sequenceCondition = new SequenceCondition(2);
-        DiscountPolicy discountPolicy = new PercentDiscountPolicy(0.1, sequenceCondition);
+        DefaultDiscountPolicy discountPolicy = new PercentDiscountPolicy(0.1, sequenceCondition);
         Movie movie = new Movie("타이타닉", Duration.ofMinutes(100), Money.wons(10000), discountPolicy);
         Screening screening = new Screening(movie, 2, LocalDateTime.of(2025, 4, 21, 10, 0));
